Split into more files for norm checks, added extra free in child if execve failed.
